requirements.txt
flask==3.1.0 #microframework web para el webservice
mysql-connector-python==9.3.0 #para hacer registros en la bd mysql
gunicorn #para crear servidor web